+++ sdk/core/azure-core-amqp/inc/azure/core/amqp/internal/message_receiver.hpp
@@ -3,6 +3,7 @@
 
 #pragma once
 
+#include "azure/core/amqp/internal/amqp_settle_mode.hpp"
 #include "azure/core/amqp/internal/models/amqp_error.hpp"
 #include "azure/core/amqp/models/amqp_message.hpp"
 #include "azure/core/amqp/models/amqp_value.hpp"
@@ -33,12 +34,6 @@ namespace Azure { namespace Core { namespace Amqp { namespace _internal {
     Error,
   };
   std::ostream& operator<<(std::ostream& stream, _internal::MessageReceiverState const& state);
-
-  enum class ReceiverSettleMode
-  {
-    First,
-    Second,
-  };
 
   class MessageReceiver;
 
@@ -104,7 +99,10 @@ namespace Azure { namespace Core { namespace Amqp { namespace _internal {
         MessageReceiver const& receiver,
         std::shared_ptr<Models::AmqpMessage> const& message)
         = 0;
-    virtual void OnMessageReceiverDisconnected(Models::_internal::AmqpError const& error) = 0;
+    virtual void OnMessageReceiverDisconnected(
+        MessageReceiver const& receiver,
+        Models::_internal::AmqpError const& error)
+        = 0;
   };
 
   /** @brief MessageReceiver