+++ sdk/core/azure-core-amqp/src/amqp/management.cpp
@@ -204,13 +204,19 @@ namespace Azure { namespace Core { namespace Amqp { namespace _detail {
     SetState(ManagementState::Closing);
     if (m_messageSender && m_messageSenderOpen)
     {
-      Log::Stream(Logger::Level::Verbose) << "ManagementClient::Close Sender" << std::endl;
+      if (m_options.EnableTrace)
+      {
+        Log::Stream(Logger::Level::Verbose) << "ManagementClient::Close Sender" << std::endl;
+      }
       m_messageSender->Close(context);
       m_messageSenderOpen = false;
     }
     if (m_messageReceiver && m_messageReceiverOpen)
     {
-      Log::Stream(Logger::Level::Verbose) << "ManagementClient::Close Receiver" << std::endl;
+      if (m_options.EnableTrace)
+      {
+        Log::Stream(Logger::Level::Verbose) << "ManagementClient::Close Receiver" << std::endl;
+      }
       m_messageReceiver->Close(context);
       m_messageReceiverOpen = false;
     }
@@ -224,8 +230,11 @@ namespace Azure { namespace Core { namespace Amqp { namespace _detail {
   {
     if (newState == oldState)
     {
-      Log::Stream(Logger::Level::Verbose)
-          << "ManagementClient::OnMessageSenderStateChanged: newState == oldState" << std::endl;
+      if (m_options.EnableTrace)
+      {
+        Log::Stream(Logger::Level::Verbose)
+            << "ManagementClient::OnMessageSenderStateChanged: newState == oldState" << std::endl;
+      }
       return;
     }
 
@@ -319,7 +328,9 @@ namespace Azure { namespace Core { namespace Amqp { namespace _detail {
     }
   }
 
-  void ManagementClientImpl::OnMessageSenderDisconnected(Models::_internal::AmqpError const& error)
+  void ManagementClientImpl::OnMessageSenderDisconnected(
+      _internal::MessageSender const&,
+      Models::_internal::AmqpError const& error)
   {
     if (error)
     {
@@ -553,6 +564,7 @@ namespace Azure { namespace Core { namespace Amqp { namespace _detail {
   }
 
   void ManagementClientImpl::OnMessageReceiverDisconnected(
+      _internal::MessageReceiver const&,
       Models::_internal::AmqpError const& error)
   {
     if (error)