+++ sdk/core/azure-core-amqp/src/amqp/private/connection_impl.hpp
@@ -58,7 +58,8 @@ namespace Azure { namespace Core { namespace Amqp { namespace _detail {
     ConnectionImpl(
         std::shared_ptr<Network::_detail::TransportImpl> transport,
         _internal::ConnectionOptions const& options,
-        _internal::ConnectionEvents* eventHandler);
+        _internal::ConnectionEvents* eventHandler,
+        _internal::ConnectionEndpointEvents* endpointEvents);
 
     ConnectionImpl(
         std::string const& hostName,
@@ -90,8 +91,6 @@ namespace Azure { namespace Core { namespace Amqp { namespace _detail {
 
     void Open();
     void Listen();
-
-    void Close();
 
     void Close(
         std::string const& condition = {},
@@ -144,6 +143,7 @@ namespace Azure { namespace Core { namespace Amqp { namespace _detail {
     Azure::Core::Amqp::Common::_internal::AsyncOperationQueue<std::unique_ptr<_internal::Session>>
         m_newSessionQueue;
     _internal::ConnectionEvents* m_eventHandler{};
+    _internal::ConnectionEndpointEvents* m_endpointEvents{};
     _internal::ConnectionState m_connectionState = _internal::ConnectionState::Start;
 
     LockType m_amqpMutex;