+++ sdk/core/azure-core-amqp/src/models/message_source.cpp
@@ -7,6 +7,7 @@
 #include "private/value_impl.hpp"
 
 #include <azure_uamqp_c/amqp_definitions_fields.h>
+
 #include <azure_uamqp_c/amqp_definitions_terminus_durability.h>
 #include <azure_uamqp_c/amqp_definitions_terminus_expiry_policy.h>