+++ sdk/core/azure-core-amqp/test/ut/management_tests.cpp
@@ -103,7 +103,8 @@ namespace Azure { namespace Core { namespace Amqp { namespace Tests {
         }
         return MockServiceEndpoint::OnMessageReceived(receiver, incomingMessage);
       }
-      void MessageReceived(std::shared_ptr<AmqpMessage> const& incomingMessage) override
+      void MessageReceived(std::string const&, std::shared_ptr<AmqpMessage> const& incomingMessage)
+          override
       {
         if (incomingMessage->ApplicationProperties.at("operation") == "Test")
         {
@@ -234,11 +235,13 @@ namespace Azure { namespace Core { namespace Amqp { namespace Tests {
       }
 
     private:
-      void MessageReceived(std::shared_ptr<AmqpMessage> const& incomingMessage) override
+      void MessageReceived(
+          std::string const& linkName,
+          std::shared_ptr<Azure::Core::Amqp::Models::AmqpMessage> const& message) override
       {
-        // Do nothing.
-        GTEST_LOG_(INFO) << "NullResponseManagementServiceEndpoint::MessageReceived("
-                         << incomingMessage << ") called.";
+        GTEST_LOG_(INFO)
+            << "NullResponseManagementServiceEndpoint::MessageReceived received on link "
+            << linkName << ": " << *message;
       }
     };
   } // namespace